requirements.txt
alembic==1.8.0
aniso8601==9.0.1
autopep8==2.0.1
Babel==2.11.0
blinker==1.5
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
dominate==2.6.0
email-validator==1.2.1
flake8==6.0.0
Flask==2.1.1
flask-babel==3.0.1
Flask-Bootstrap==3.3.7.1
Flask-Bootstrap4==4.0.2
Flask-Datepicker==0.14
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Table==0.5.0
Flask-WTF==1.0.1
greenlet==2.0.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
Mako==1.2.1
MarkupSafe==2.1.1
mccabe==0.7.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyMySQL==1.0.2
python-dotenv==0.20.0
pytz==2022.7.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
SQLAlchemy-Utils==0.38.3
tomli==2.0.1
urllib3==1.26.9
visitor==0.1.3
waitress==2.1.2
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0
